refactor(onboarding): extract startup config and profile defaults
