+++ crates/tau-coding-agent/src/commands.rs
@@ -2,6 +2,10 @@ use super::*;
 use crate::extension_manifest::{
     dispatch_extension_registered_command, ExtensionRegisteredCommandAction,
 };
+#[cfg(test)]
+use crate::runtime_types::{
+    ProfileAuthDefaults, ProfileMcpDefaults, ProfilePolicyDefaults, ProfileSessionDefaults,
+};
 use tau_session::{
     execute_session_diff_command, execute_session_search_command, execute_session_stats_command,
     parse_session_diff_args, parse_session_stats_args,
