+++ crates/tau-coding-agent/src/runtime_types.rs
@@ -1,11 +1,15 @@
 use std::path::PathBuf;
 
 use crate::extension_manifest::ExtensionRegisteredCommand;
-use crate::{default_provider_auth_method, Cli, ModelCatalog, ProviderAuthMethod};
-use serde::{Deserialize, Serialize};
+use crate::{Cli, ModelCatalog};
 pub(crate) use tau_diagnostics::DoctorCommandConfig;
 #[cfg(test)]
 pub(crate) use tau_diagnostics::{DoctorMultiChannelReadinessConfig, DoctorProviderKeyStatus};
+pub(crate) use tau_onboarding::startup_config::ProfileDefaults;
+#[cfg(test)]
+pub(crate) use tau_onboarding::startup_config::{
+    ProfileAuthDefaults, ProfileMcpDefaults, ProfilePolicyDefaults, ProfileSessionDefaults,
+};
 pub(crate) use tau_provider::AuthCommandConfig;
 use tau_session::SessionImportMode;
 
@@ -15,82 +19,6 @@ pub(crate) struct SkillsSyncCommandConfig {
     pub(crate) default_lock_path: PathBuf,
     pub(crate) default_trust_root_path: Option<PathBuf>,
     pub(crate) doctor_config: DoctorCommandConfig,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub(crate) struct ProfileSessionDefaults {
-    pub(crate) enabled: bool,
-    pub(crate) path: Option<String>,
-    pub(crate) import_mode: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub(crate) struct ProfilePolicyDefaults {
-    pub(crate) tool_policy_preset: String,
-    pub(crate) bash_profile: String,
-    pub(crate) bash_dry_run: bool,
-    pub(crate) os_sandbox_mode: String,
-    pub(crate) enforce_regular_files: bool,
-    pub(crate) bash_timeout_ms: u64,
-    pub(crate) max_command_length: usize,
-    pub(crate) max_tool_output_bytes: usize,
-    pub(crate) max_file_read_bytes: usize,
-    pub(crate) max_file_write_bytes: usize,
-    pub(crate) allow_command_newlines: bool,
-}
-
-fn default_profile_mcp_context_providers() -> Vec<String> {
-    vec![
-        "session".to_string(),
-        "skills".to_string(),
-        "channel-store".to_string(),
-    ]
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub(crate) struct ProfileMcpDefaults {
-    #[serde(default = "default_profile_mcp_context_providers")]
-    pub(crate) context_providers: Vec<String>,
-}
-
-impl Default for ProfileMcpDefaults {
-    fn default() -> Self {
-        Self {
-            context_providers: default_profile_mcp_context_providers(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub(crate) struct ProfileAuthDefaults {
-    #[serde(default = "default_provider_auth_method")]
-    pub(crate) openai: ProviderAuthMethod,
-    #[serde(default = "default_provider_auth_method")]
-    pub(crate) anthropic: ProviderAuthMethod,
-    #[serde(default = "default_provider_auth_method")]
-    pub(crate) google: ProviderAuthMethod,
-}
-
-impl Default for ProfileAuthDefaults {
-    fn default() -> Self {
-        Self {
-            openai: default_provider_auth_method(),
-            anthropic: default_provider_auth_method(),
-            google: default_provider_auth_method(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub(crate) struct ProfileDefaults {
-    pub(crate) model: String,
-    pub(crate) fallback_models: Vec<String>,
-    pub(crate) session: ProfileSessionDefaults,
-    pub(crate) policy: ProfilePolicyDefaults,
-    #[serde(default)]
-    pub(crate) mcp: ProfileMcpDefaults,
-    #[serde(default)]
-    pub(crate) auth: ProfileAuthDefaults,
 }
 
 #[derive(Debug, Clone, Copy)]
