+++ crates/tau-onboarding/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod startup_config;
 pub mod startup_model_resolution;
 pub mod startup_policy;
 pub mod startup_prompt_composition;
